transaction model; greenhouse; in progress

+++ robnur/shared/src/main/java/at/ac/tuwien/complang/vpsbcm/robnur/shared/plants/CultivationInformation.java
@@ -7,7 +7,7 @@ public abstract class CultivationInformation implements Serializable {
     /* 0.2 ... 0.5 */
     private float growthRate;
 
-    private float harvest;  // number of vegetables or flowers which can be harvested
+    private int harvest;  // number of vegetables or flowers which can be harvested
 
     private int upgradeLevel;
 
@@ -41,11 +41,11 @@ public abstract class CultivationInformation implements Serializable {
         this.growthRate = growthRate;
     }
 
-    public float getHarvest() {
+    public int getHarvest() {
         return harvest;
     }
 
-    public void setHarvest(float harvest) {
+    public void setHarvest(int harvest) {
         this.harvest = harvest;
     }
